requirements.txt
aiosmtpd==1.4.2
alembic==1.9.1
apt-clone==0.2.1
apturl==0.5.2
asn1crypto==0.24.0
atpublic==3.1.1
attrs==22.1.0
Babel==2.11.0
bcrypt==4.0.1
beautifulsoup4==4.10.0
blinker==1.5
Bootstrap-Flask==2.2.0
Brlapi==0.8.3
catfish==4.16.3
ccsm==0.9.14.1
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
command-not-found==0.3
compizconfig-python==0.9.14.1
configobj==5.0.6
cryptography==39.0.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
dnspython==2.2.1
dominate==2.7.0
email-validator==1.3.0
enum34==1.1.6
eyeD3==0.8.10
filelock==3.6.0
Flask==2.2.2
Flask-Babel==2.0.0
Flask-Bcrypt==1.0.1
flask-change-password==0.0.7
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
Flask-Moment==1.0.5
Flask-MySQL==1.5.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
grpcio==1.30.2
httpie==2.6.0
httplib2==0.20.2
idna==3.4
ifaddr==0.1.7
imageio==2.28.1
IMDbPY==2021.4.18
importlib-metadata==4.6.4
ipaddress==1.0.22
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
langdetect==1.0.9
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.2
louis==3.20.0
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==2.17.0
marshmallow-sqlalchemy==0.15.0
menulibre==2.2.2
more-itertools==8.10.0
mugshot==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.32
netaddr==0.8.0
netifaces==0.11.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.0
onboard==1.4.1
onetimepass==1.0.1
packaging==22.0
passlib==1.7.1
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
Pygments==2.11.2
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.6.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
PyQRCode==1.2.1
PyQt5==5.15.6
PyQt5-sip==12.9.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.7.5
python-debian===0.1.43ubuntu1
python-dotenv==0.21.0
python-magic==0.4.24
python-xapp==2.2.2
pytz==2022.6
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
scikit-image==0.20.0
scipy==1.10.1
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.1
speaklater==1.3
SQLAlchemy==1.4.45
SQLAlchemy-serializer==1.4.1
systemd-python==234
tifffile==2023.4.12
tldextract==3.1.2
typing_extensions==4.5.0
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.13
visitor==0.1.3
wadllib==1.3.6
Werkzeug==2.2.2
WTForms==3.0.1
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
zipp==1.0.0
